Find service cycles without spending heap memory
Instead of recording the services being requested
in an extra queue, throw an exception on a cycle
instead and enrich it as we go back up the stack.

This completely removes the cost of cycle detection
during normal operation while still giving us proper
error reporting when something goes wrong.