Remove unneccessary objects from service registry
This minimizes the memory footprint of service registries
with the following improvements:

The isolation beteen parent and child is now handled by
explicitly checking whether a service came from the same
registry instead of wrapping services in an isolation layer.

The lookup cache has been removed, as the lookup algorithm
is now cheap enough to work without a cache.

The type spec concept has been removed and replaced by a
static "isSatisfiedBy(Type, Type)" method. The cache for the
specs was also removed as a result.

An array of services is only created if there are multiple
implementations for an interface.

Looking up Object.class is no longer supported and it is
no longer tracked as a potential service class.